revert keypoint refactoring. (#869)
* revert keypoint refactoring.

* fix test.

* fix test.

* fix test.
