+++ keras_cv/keypoint/__init__.py
@@ -12,10 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from keras_cv.ops.point_cloud import _box_area
-from keras_cv.ops.point_cloud import _center_xyzWHD_to_corner_xyz
-from keras_cv.ops.point_cloud import _is_on_lefthand_side
-from keras_cv.ops.point_cloud import coordinate_transform
-from keras_cv.ops.point_cloud import is_within_box2d
-from keras_cv.ops.point_cloud import is_within_box3d
-from keras_cv.ops.point_cloud import spherical_coordinate_transform
+from keras_cv.keypoint.converters import convert_format
+from keras_cv.keypoint.formats import REL_XY
+from keras_cv.keypoint.formats import XY
+from keras_cv.keypoint.utils import filter_out_of_image
